# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "LibreCov is a self-hosted open-source code coverage history viewer. This API is mostly compatible with the Coveralls API.",
        "title": "LibreCov API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "LibreCov Support",
            "url": "https://github.com/Frantche/Librecov",
            "email": "support@librecov.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/Frantche/Librecov/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/",
    "paths": {
...
        "/auth/login": {
            "get": {
                "description": "Redirects to OIDC provider for authentication",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Initiate OIDC login",
                "responses": {
                    "302": {
                        "description": "Redirect to OIDC provider",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get information about the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Get current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Frantche_Librecov_backend_internal_models.User"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_Frantche_Librecov_backend_internal_models.Build": {
            "type": "object",
            "properties": {
                "branch": {
                    "type": "string"
                },
                "build_num": {
                    "type": "integer"
                },
                "commit_msg": {
                    "type": "string"
                },
                "commit_sha": {
                    "type": "string"
                },
                "coverage_rate": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "jobs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Frantche_Librecov_backend_internal_models.Job"
                    }
                },
                "project": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_Frantche_Librecov_backend_internal_models.Project"
                        }
                    ]
                },
                "project_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_Frantche_Librecov_backend_internal_models.Job": {
            "type": "object",
            "properties": {
                "build": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_Frantche_Librecov_backend_internal_models.Build"
                        }
                    ]
                },
                "build_id": {
                    "type": "integer"
                },
                "coverage_rate": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "data": {
                    "description": "JSON data",
                    "type": "string"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Frantche_Librecov_backend_internal_models.JobFile"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "job_number": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_Frantche_Librecov_backend_internal_models.JobFile": {
            "type": "object",
            "properties": {
                "coverage": {
                    "description": "JSON array",
                    "type": "string"
                },
                "coverage_rate": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "job": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_Frantche_Librecov_backend_internal_models.Job"
                        }
                    ]
                },
                "job_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_Frantche_Librecov_backend_internal_models.Project": {
            "type": "object",
            "properties": {
                "base_url": {
                    "type": "string"
                },
                "builds": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Frantche_Librecov_backend_internal_models.Build"
                    }
                },
                "coverage_rate": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "current_branch": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "shares": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Frantche_Librecov_backend_internal_models.ProjectShare"
                    }
                },
                "token": {
                    "type": "string"
                },
                "tokens": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Frantche_Librecov_backend_internal_models.ProjectToken"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_Frantche_Librecov_backend_internal_models.User"
                        }
                    ]
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_Frantche_Librecov_backend_internal_models.ProjectShare": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "group_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "project_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_Frantche_Librecov_backend_internal_models.ProjectToken": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_used": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_Frantche_Librecov_backend_internal_models.User": {
            "type": "object",
            "properties": {
                "admin": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "email_verified": {
                    "type": "boolean"
                },
                "groups": {
                    "description": "JSON array of group names from OIDC token",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "projects": {
                    "description": "Relationships",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Frantche_Librecov_backend_internal_models.Project"
                    }
                },
                "token": {
                    "type": "string"
                },
                "tokens": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Frantche_Librecov_backend_internal_models.UserToken"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_Frantche_Librecov_backend_internal_models.UserToken": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_used": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ProjectToken": {
            "description": "Project token for API access",
            "type": "apiKey",
            "name": "token",
            "in": "query"
        }
    }
}